fix: separada rota de criacao de usuario das demais
