+++ src/middlewares/ensure-authenticated.ts
@@ -30,6 +30,8 @@ req.user = {
     user_id
 }
 
+console.log(req.user)
+
 return next()
 
     } catch (error) {
