+++ src/routes/users-routes.ts
@@ -5,7 +5,6 @@ const usersRoutes = Router();
 
 const usersController = new UsersController
 
-usersRoutes.post("/", usersController.create)
 usersRoutes.get("/", usersController.index)
 usersRoutes.put("/:id", usersController.update)
 usersRoutes.delete("/:id", usersController.delete)
